+++ HISTORY.rst
@@ -19,6 +19,7 @@ Unreleased
 * ALCN-812: Add missing ir.model.access
 * Delivery rounds: Fix delivery in case of lot
 * Sales order confirmation if no stock: fix move assignment in case of no quant available
+* ALCN-810: Allow to receive a lot that already exist: change reception picking type to use existing lot instead of create new lot
 
 **Build**